[refactor] make backend a package
